+++ CHANGELOG.md
@@ -3,6 +3,16 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 
+v1.1.0
+### Added
+- Database
+- EF core
+- Reminder service for saving reminders
+- Command for saving reminders
+- Help for reminders
+### Changed
+### Fixed
+
 v1.0.3
 ### Added
 ### Changed